+++ public/src/main/java/com/revenuecat/purchases/models/PurchaseDetails.kt
@@ -9,31 +9,77 @@ import org.json.JSONObject
 
 @Parcelize
 @TypeParceler<JSONObject, JSONObjectParceler>()
-// TODO: docs
 data class PurchaseDetails(
+    /*
+     * Unique Google order identifier for the purchased transaction.
+     *
+     * Only available for non-restored Google purchases.
+     */
     val orderId: String?,
+
+    /*
+     * Product Id.
+     */
     val sku: String,
+
+    /*
+     * Type of the product associated with the purchase.
+     */
     val type: ProductType,
+
+    /*
+     * Time the product was purchased, in milliseconds since the epoch.
+     */
     val purchaseTime: Long,
+
+    /*
+     * Token that uniquely identifies a purchase.
+     * For Amazon purchases, this is equivalent to the receipt ID.
+     */
     val purchaseToken: String,
+
+    /*
+     * State of the purchase.
+     */
     val purchaseState: RevenueCatPurchaseState,
 
     /*
-     * Null for restored purchases
+     * Whether the subscription renews automatically.
+     *
+     * Null for Google restored purchases.
      */
     val isAutoRenewing: Boolean?,
 
     /*
-     * Null for Amazon
+     * String containing the signature of the Google purchase data that was signed with the private key of
+     * the developer.
+     *
+     * Null for Amazon.
      */
     val signature: String?,
 
+    /*
+     * Returns a JSONObject format that contains details about the purchase.
+     */
+    @SuppressWarnings("ForbiddenComment")
+    // TODO: change to String
     val originalJson: JSONObject,
 
+    /*
+     * Offering that was presented when making the purchase. Not available for restored purchases.
+     */
     val presentedOfferingIdentifier: String?,
 
+    /*
+     * Amazon unique identifier of the customer.
+     *
+     * Null for Google
+     */
     val storeUserID: String?,
 
+    /*
+     * One of [PurchaseType] indicating the type of purchase.
+     */
     val purchaseType: PurchaseType
 ) : Parcelable
